requirements.txt
Flask
pytesseract
Pillow
googletrans==4.0.0-rc1
sumy
PyPDF2
pdf2image
gunicorn
PyMuPDF
